+++ SystemGeneration/Matrix.h
@@ -58,7 +58,7 @@ namespace matrixes
 
 		~Matrix()
 		{
-			for (int i = 0; i < data.size(); i++)
+			for (size_t i = 0; i < data.size(); i++)
 				delete data[i];
 
 			data.clear();
