+++ SystemGeneration/Polynomial.h
@@ -30,7 +30,7 @@ namespace polynomials {
 		// для тестирования
 		Polynomial(const std::vector <std::vector <int>> input) : n_max(0)
 		{
-			for (int i = 0; i < input.size(); i++)
+			for (size_t i = 0; i < input.size(); i++)
 			{
 				std::vector <int> cur = input[i];
 				terms.push_back(Monomial(cur));
